refactor to support calls through API
